adding repository classes and data script - configure h2 and mysql databases

+++ src/main/java/me/mepv/hotel/entity/Room.java
@@ -6,6 +6,7 @@ import lombok.NoArgsConstructor;
 import lombok.Setter;
 
 import javax.persistence.*;
+import java.math.BigDecimal;
 
 @Getter
 @Setter
@@ -18,7 +19,7 @@ public class Room extends BaseEntity {
     @Column(name = "room_number")
     private Integer roomNumber;
     private Integer floor;
-    private Double price;
+    private BigDecimal price;
     private Boolean booked = false;
     @ManyToOne(cascade = CascadeType.ALL, fetch = FetchType.LAZY)
     @JoinColumn(name = "room_type_id")
